CCPP framework (metadata parser) update, and update several units in CCPP metadata in fv3atm and ccpp-physics, use Python 3.6+ (#907)
With the metadata parser update in NCAR/ccpp-framework#415, a stricter checking of units will be enforced. This includes the replacement of "convenience units" such as log(Pa) with 1. Corresponding changes are made in ccpp-physics and fv3atm.

The ccpp-framework update requires Python 3.6 or newer - load Python 3.7.9 modules on Cheyenne, all other platforms are already using 3.6 or newer.