Added pointers to merged nml and added export_tiled feature
Variables in the merged nml are now all represented with pointers. These
can be edited by users or will use the default values in default_vars.sh.
A new export_tiled section has been added to default_vars.sh for configurations
that use a tile set that is not default (C96 is set to default).