Turn on old tests and remove control_csawmgt
Tests that still use the GFSv16 core are turned back on. Work to
update these tests to be compatible with GFSv17 has started. The
tests will be updated as the work on the individual tests are
completed. control_csawmgt and its associated files have been
removed under the recommendation by the appropriate team.