Add write_restart_at_endofrun option to ufs.configure
There were multiple regression tests that did not have the
write_restart_at_endofrun option in the ufs.configure file.
Now that the write_restart_at_endofrun has been fixed in MOM6
and CICE, this option will be added to the ufs.configure files.
This does not cause change the RT baselines.