baseline validation for ww3 3b038ff6
* tested all wave-containing tests with modifications for restart
file naming to allow for the custom filenaming of binary restarts.
This feature is present in the current WW3 code, but will be removed
once we enable netcdf restarts. Temporary code was added to allow the
binary restart to have the existing format of casename+ww3.r+timestring.
With this modification, all baselines were B4B.