Megre global_control_lnd_iau into global_control
This allows the control_c48_lnd_iau test to be run
off of the global_control.nml.IN namelist and adds
the Land IAU defaults to default_vars.sh

Even though PERT_CLDS was set in the original test config,
the namlist did not have this so the option was not turned on.
The test settings reflect the original version of the test and
no baseline changes have been made.